refactor: nettoyer et modulariser utils.py (717 → 34 lignes)
- Créé config.py : configuration et constantes (64 lignes)
- Créé metrics.py : calcul WER et nettoyage de texte (113 lignes)
- Créé api_clients.py : clients OpenRouter/Transkribus (334 lignes)
- Créé reporting.py : génération de rapports Markdown (188 lignes)
- Simplifié utils.py : wrapper pour compatibilité (34 lignes)
- Supprimé code mort : fonction copy_random_images() non utilisée
- Maintenu 100% de compatibilité avec le code existant

Résultat : code mieux organisé, plus maintenable et modulaire
